gcc porting fixes

+++ include/ofp/yaml/yactiontype.h
@@ -22,7 +22,7 @@ struct ScalarTraits<ofp::ActionType> {
 
   static StringRef input(StringRef scalar, void *ctxt, ofp::ActionType &value) {
     if (!value.parse(scalar)) {
-      uint64_t num;
+      unsigned long long num;
       if (llvm::getAsUnsignedInteger(scalar, 0, num) == 0) {
         value.setNative(ofp::UInt16_narrow_cast(num));
         return "";
